Add option to ignore certificate verification
An optional section 'options' is added to the config file which can be
used to set global options within the client - the first of which
implemented is the boolean flag able to be set to ignore certificate
verification within the http request. This is to allow interaction with
the staging server which has a self-signed ssl certificate.