Implement empty targets and resolution of filepaths
The target can now be left empty and will be passed through to the NLDS
as such, resulting in slightly different behaviour. No target will
extract the files back to their original location, a given target will
allow the files ot be placed somewhere else - this will be improved by
future inclusion of common_path storage.

Given filepaths in the no-target case will be resolved fully, the target
is also reolved fully in the other case.